move PayloadEnvelopeV3 to e2e utils helper traits
