+++ crates/e2e-test-utils/src/node.rs
@@ -1,4 +1,7 @@
-use crate::{engine_api::EngineApiHelper, network::NetworkHelper, payload::PayloadHelper};
+use crate::{
+    engine_api::EngineApiHelper, network::NetworkHelper, payload::PayloadHelper,
+    traits::PayloadEnvelopeV3,
+};
 use alloy_rpc_types::BlockNumberOrTag;
 use eyre::Ok;
 use reth::{
@@ -58,7 +61,7 @@ where
         attributes_generator: impl Fn(u64) -> EngineType::PayloadBuilderAttributes,
     ) -> eyre::Result<(B256, B256)>
     where
-        EngineType::ExecutionPayloadV3: From<EngineType::BuiltPayload>,
+        EngineType::ExecutionPayloadV3: From<EngineType::BuiltPayload> + PayloadEnvelopeV3,
     {
         // push tx into pool via RPC server
         let tx_hash = self.inject_tx(raw_tx).await?;
