+++ crates/engine-primitives/src/traits.rs
@@ -6,11 +6,8 @@ use reth_primitives::{
     Address, ChainSpec, Header, SealedBlock, Withdrawals, B256, U256,
 };
 use reth_rpc_types::{
-    engine::{
-        ExecutionPayloadEnvelopeV3, OptimismExecutionPayloadEnvelopeV3, OptimismPayloadAttributes,
-        PayloadAttributes as EthPayloadAttributes, PayloadId,
-    },
-    ExecutionPayloadV3, Withdrawal,
+    engine::{OptimismPayloadAttributes, PayloadAttributes as EthPayloadAttributes, PayloadId},
+    Withdrawal,
 };
 
 /// Represents a built payload type that contains a built [SealedBlock] and can be converted into
@@ -155,23 +152,5 @@ impl PayloadAttributes for OptimismPayloadAttributes {
         }
 
         Ok(())
-    }
-}
-
-/// The execution payload envelope type.
-pub trait PayloadEnvelopeV3: Send + Sync + std::fmt::Debug {
-    /// Returns execution payload from envelope
-    fn execution_payload(&self) -> ExecutionPayloadV3;
-}
-
-impl PayloadEnvelopeV3 for OptimismExecutionPayloadEnvelopeV3 {
-    fn execution_payload(&self) -> ExecutionPayloadV3 {
-        self.execution_payload.clone()
-    }
-}
-
-impl PayloadEnvelopeV3 for ExecutionPayloadEnvelopeV3 {
-    fn execution_payload(&self) -> ExecutionPayloadV3 {
-        self.execution_payload.clone()
     }
 }
